Move JsonReader to log4j-core
