+++ log4j-core/src/main/java/org/apache/logging/log4j/core/util/JsonReader.java
@@ -14,7 +14,7 @@
  * See the license for the specific language governing permissions and
  * limitations under the license.
  */
-package org.apache.logging.log4j.util;
+package org.apache.logging.log4j.core.util;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
